+++ src/main/java/module-info.java
@@ -10,6 +10,7 @@ module org.app {
     requires org.kordamp.ikonli.javafx;
     requires org.kordamp.ikonli.core;
     requires org.kordamp.ikonli.material2;
+    requires org.slf4j;
 
     exports org.app;
     exports org.controller;
@@ -24,8 +25,10 @@ module org.app {
     exports org.model.goal;
     exports org.context;
     exports org.manager;
+    exports org.logging;
     opens org.app to javafx.fxml;
     opens org.controller to javafx.fxml;
     opens org.utilities to javafx.fxml;
     opens org.service to com.opencsv;
+    opens org.logging to javafx.fxml;
 }