+++ src/main/java/org/controller/AddTransactionDialog.java
@@ -12,6 +12,8 @@ import org.manager.DbManager;
 import org.model.Formation;
 import org.model.symbol.Symbol;
 import org.model.transaction.Transaction;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -28,6 +30,8 @@ import static org.utilities.Utilities.*;
 
 
 public class AddTransactionDialog implements Initializable {
+    private static final Logger log = LoggerFactory.getLogger(AddTransactionDialog.class);
+
     @FXML
     public Button save;
     @FXML
@@ -177,13 +181,8 @@ public class AddTransactionDialog implements Initializable {
                 db.closeBdConnection();
                 mainController.hideModal();
                 resetForm();
-                System.out.println("Transaction added successfully!!");
-            } catch (SQLException | IOException e) {
-                System.out.println("Saving Transaction Failed");
-                throw new RuntimeException(e);
+            } catch (SQLException | IOException ignored) {
             }
-        } else {
-            System.out.println("Form validation failed");
         }
     }
 
@@ -215,9 +214,11 @@ public class AddTransactionDialog implements Initializable {
             possibleProfitTicks.pseudoClassStateChanged(Styles.STATE_DANGER, isPossibleProfitTicksError);
             possibleLossTicks.pseudoClassStateChanged(Styles.STATE_DANGER, isPossibleLossTicksError);
             timePeriod.pseudoClassStateChanged(Styles.STATE_DANGER, isTimePeriodError);
+            log.error("Add transaction form validation failed");
             return false;
         }
 
+        log.info("Add transaction form validation successful");
         return true;
     }