+++ src/main/java/org/controller/DailyPrepController.java
@@ -19,6 +19,7 @@ import org.manager.DbManager;
 import org.model.dailyPrep.DailyPrep;
 import org.model.dailyPrep.DailyPrepItems;
 import org.utilities.DateCellTreeTable;
+import org.utilities.SaveHandler;
 
 import java.io.IOException;
 import java.net.URL;
@@ -28,7 +29,7 @@ import java.util.ResourceBundle;
 
 import static org.utilities.Utilities.*;
 
-public class DailyPrepController extends Pane implements Initializable {
+public class DailyPrepController extends Pane implements Initializable, SaveHandler {
 
     @FXML
     public TreeTableView<Object> tableView;
@@ -73,6 +74,7 @@ public class DailyPrepController extends Pane implements Initializable {
         ControllerRegistry.register(DailyPrepController.class, this);
         this.mainController = ControllerRegistry.get(MainController.class);
         saveBtn.getStyleClass().add(Styles.ACCENT);
+        saveBtn.setOnAction(event -> save());
 
         ///Populate list
         dateColumn.setCellValueFactory(new TreeItemPropertyValueFactory<>("date"));
@@ -133,7 +135,8 @@ public class DailyPrepController extends Pane implements Initializable {
     }
 
     @FXML
-    public void saveForm() {
+    @Override
+    public void save() {
         selectedSymbol.setDailyEvents(dailyEvents.getText());
         selectedSymbol.setHourlyTrend(hourlyTrend.getText());
         selectedSymbol.setHalfHourlyTrend(halfHourlyTrend.getText());
@@ -150,9 +153,7 @@ public class DailyPrepController extends Pane implements Initializable {
             db.addDailyPrepItem(selectedSymbol);
             GlobalContext.getDailyPrep().replaceMaster(db.getAllDailyPrepData());
             db.closeBdConnection();
-            System.out.println("Symbol updated successfully!!");
-        } catch (IOException | SQLException e) {
-            throw new RuntimeException(e);
+        } catch (IOException | SQLException ignored) {
         }
     }
 
@@ -199,9 +200,7 @@ public class DailyPrepController extends Pane implements Initializable {
                 db.deleteSymbol(selectedSymbol.getDailyPrepId());
                 GlobalContext.getDailyPrep().replaceMaster(db.getAllDailyPrepData());
                 db.closeBdConnection();
-                System.out.println("Symbol deleted successfully!!");
-            } catch (IOException | SQLException e) {
-                throw new RuntimeException(e);
+            } catch (IOException | SQLException ignored) {
             }
         }
     }
@@ -217,9 +216,7 @@ public class DailyPrepController extends Pane implements Initializable {
                 db.deleteDay(selectedSymbol.getDailyPrepDateId());
                 GlobalContext.getDailyPrep().replaceMaster(db.getAllDailyPrepData());
                 db.closeBdConnection();
-                System.out.println("Day deleted successfully!!");
-            } catch (IOException | SQLException e) {
-                throw new RuntimeException(e);
+            } catch (IOException | SQLException ignored) {
             }
         }
     }