+++ src/main/java/org/controller/MenuBarController.java
@@ -15,7 +15,10 @@ import org.manager.DbManager;
 import org.model.account.Account;
 import org.model.symbol.Symbol;
 import org.model.transaction.Transaction;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.utilities.AppProperties;
+import org.logging.PopoutLogger;
 
 import java.io.File;
 import java.io.IOException;
@@ -30,6 +33,7 @@ import static org.service.csvReader.*;
 import static org.utilities.Utilities.closeApp;
 
 public class MenuBarController extends VBox implements Initializable {
+    private static final Logger log = LoggerFactory.getLogger(MenuBarController.class);
 
     @FXML
     MenuBar menuBar;
@@ -72,15 +76,14 @@ public class MenuBarController extends VBox implements Initializable {
             List<Transaction> importedTransactions = getAllTransactions(file);
             try {
                 db.setBdConnection();
-            } catch (IOException e) {
-                System.out.println("Failed to connect to DB");
-                throw new RuntimeException(e);
+            } catch (IOException ignored) {
             }
             for (Transaction tran : importedTransactions) {
                 db.addTransaction(tran);
             }
             GlobalContext.getTransactions().replaceMaster(db.getAllTransactions());
             db.closeBdConnection();
+            log.info("File {} imported successfully", file);
             fileSuccessAlert.setContentText("File successfully imported!");
             fileSuccessAlert.showAndWait();
         }
@@ -91,6 +94,7 @@ public class MenuBarController extends VBox implements Initializable {
         String path = csvDirectorySelector("Transactions");
         if (path != null) {
             writeItemsToCSV(GlobalContext.getTransactions().getMaster(), path);
+            log.info("All transactions file exported successfully");
             fileSuccessAlert.setContentText("File successfully exported!");
             fileSuccessAlert.showAndWait();
         }
@@ -101,6 +105,7 @@ public class MenuBarController extends VBox implements Initializable {
         String path = csvDirectorySelector("Transactions");
         if (path != null) {
             writeItemsToCSV(GlobalContext.getTransactions().getFiltered(), path);
+            log.info("All selected transactions file exported successfully");
             fileSuccessAlert.setContentText("File successfully exported!");
             fileSuccessAlert.showAndWait();
         }
@@ -113,7 +118,7 @@ public class MenuBarController extends VBox implements Initializable {
         if (selectedFile != null) {
             return selectedFile.getPath();
         } else {
-            System.out.println("No File Selected");
+            log.info("No file selected");
             return null;
         }
     }
@@ -130,7 +135,7 @@ public class MenuBarController extends VBox implements Initializable {
         if (selectedFile != null) {
             return selectedFile.getAbsolutePath();
         } else {
-            System.out.println("No File Selected");
+            log.info("No File Selected");
             return null;
         }
     }
@@ -176,6 +181,7 @@ public class MenuBarController extends VBox implements Initializable {
         String path = csvDirectorySelector("Symbols");
         if (path != null) {
             writeItemsToCSV(GlobalContext.getSymbols().getMaster(), path);
+            log.info("All symbols exported successfully");
             fileSuccessAlert.setContentText("File successfully exported!");
             fileSuccessAlert.showAndWait();
         }
@@ -189,6 +195,7 @@ public class MenuBarController extends VBox implements Initializable {
             List<Symbol> importedSymbols = getAllSymbols(file);
             addAllSymbol(importedSymbols);
             GlobalContext.getSymbols().setAllMaster(importedSymbols);
+            log.info("All symbols imported successfully");
             fileSuccessAlert.setContentText("File successfully imported!");
             fileSuccessAlert.showAndWait();
         }
@@ -213,6 +220,7 @@ public class MenuBarController extends VBox implements Initializable {
             List<Account> importedTransactions = getAllAccountTransactions(file);
             addAllAccountTransactions(importedTransactions);
             GlobalContext.getAccounts().setAllMaster(importedTransactions);
+            log.info("All transactions imported successfully");
             fileSuccessAlert.setContentText("File successfully imported!");
             fileSuccessAlert.showAndWait();
         }
@@ -223,10 +231,15 @@ public class MenuBarController extends VBox implements Initializable {
         String path = csvDirectorySelector("Account_Transactions");
         if (path != null) {
             writeItemsToCSV(GlobalContext.getAccounts().getMaster(), path);
+            log.info("All account transactions exported successfully");
             fileSuccessAlert.setContentText("File successfully exported!");
             fileSuccessAlert.showAndWait();
         }
 
     }
 
+    @FXML
+    void showLogs() {
+        PopoutLogger.display();
+    }
 }