+++ src/main/java/org/manager/DbManager.java
@@ -8,6 +8,8 @@ import org.model.goal.Goal;
 import org.model.journal.Journal;
 import org.model.transaction.Transaction;
 import org.service.SqliteConnection;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.sql.*;
@@ -16,14 +18,14 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class DbManager {
+    private static final Logger log = LoggerFactory.getLogger(DbManager.class);
     Connection bdConnection;
 
     public void setBdConnection() throws IOException {
         try {
             bdConnection = SqliteConnection.getConnection();
         } catch (IOException e) {
-            System.out.println("Failed to connect to DB");
-            e.printStackTrace();
+            log.error("Failed to connect to DB", e);
         }
     }
 
@@ -37,8 +39,7 @@ public class DbManager {
         try {
             return !bdConnection.isClosed();
         } catch (SQLException e) {
-            System.out.println("DB error");
-            e.printStackTrace();
+            log.error("DB error: {}", e.getMessage());
             return false;
         }
     }
@@ -52,7 +53,7 @@ public class DbManager {
                 ResultSet rs = stmt.executeQuery(sql);
                 exists = rs.next(); // If rs.next() returns true, a row was found, meaning the table exists.
             } catch (SQLException e) {
-                System.err.println("Error checking" + tableName + " existence: " + e.getMessage());
+                log.error("Error checking {} existence: {}", tableName, e.getMessage());
                 return exists;
             } finally {
                 return exists;
@@ -70,8 +71,9 @@ public class DbManager {
                 String sql = "PRAGMA foreign_keys;";
                 ResultSet rs = stmt.executeQuery(sql);
                 exists = rs.getInt(1) == 1;
+                log.info("Foreign key enabled: {}", exists);
             } catch (SQLException e) {
-                System.err.println("Error checking foreign_keys existence: " + e.getMessage());
+                log.error("Error checking foreign_keys existence: {}", e.getMessage());
                 return exists;
             } finally {
                 return exists;
@@ -105,8 +107,9 @@ public class DbManager {
                         "timePeriod TEXT NOT NULL" +
                         ");";
                 statement.execute(createTableSQL);
+                log.info("Table transactions created");
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to create new DB: {}", e.getMessage());
             } finally {
                 statement.close();
             }
@@ -120,8 +123,9 @@ public class DbManager {
                 statement = bdConnection.createStatement();
                 String createTableSQL = "PRAGMA foreign_keys = ON;";
                 statement.execute(createTableSQL);
+                log.info("Foreign key enabled");
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to enable foreign keys on new DB: {}", e.getMessage());
             } finally {
                 statement.close();
             }
@@ -138,8 +142,9 @@ public class DbManager {
                         "date ANY NOT NULL\n" +
                         ");";
                 statement.execute(createTableSQL);
+                log.info("Table DailPrepDate created");
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to create table DailyPrepDate on new DB: {}", e.getMessage());
             } finally {
                 statement.close();
             }
@@ -167,8 +172,9 @@ public class DbManager {
                         "FOREIGN KEY (dailyPrepDateId) REFERENCES parent_table(DailyPrepDate) ON DELETE CASCADE ON UPDATE NO ACTION\n" +
                         ");";
                 statement.execute(createTableSQL);
+                log.info("Table DailyPrep created");
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to create table DailyPrep on new DB: {}", e.getMessage());
             } finally {
                 statement.close();
             }
@@ -187,8 +193,9 @@ public class DbManager {
                         "symbol TEXT NOT NULL" +
                         ");";
                 statement.execute(createTableSQL);
+                log.info("Table Journal created");
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to create table Journal on new DB: {}", e.getMessage());
             } finally {
                 statement.close();
             }
@@ -208,8 +215,9 @@ public class DbManager {
                         "achieved INTEGER NOT NULL" +
                         ");";
                 statement.execute(createTableSQL);
+                log.info("Table Goal created");
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to create table Goal on new DB: {}", e.getMessage());
             } finally {
                 statement.close();
             }
@@ -247,13 +255,13 @@ public class DbManager {
                 }
                 return transactions;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Get All Transactions failed: {}", e.getMessage());
             } finally {
+                log.info("All transaction retrieved.");
                 ps.close();
                 rs.close();
             }
         }
-        System.out.println("Get All Transactions failed! DB is not connected");
         return null;
     }
 
@@ -279,8 +287,9 @@ public class DbManager {
             ps.setString(15, transaction.getTimePeriod());
             ps.executeUpdate();
             ps.close();
+            log.info("Transaction added successfully");
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Add Transaction failed: {}", e.getMessage());
         }
     }
 
@@ -307,8 +316,9 @@ public class DbManager {
             ps.setInt(16, transaction.getId());
             ps.executeUpdate();
             ps.close();
+            log.info("Transaction id:{} updated successfully", transaction.getId());
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Update Transaction failed: {}", e.getMessage());
         }
     }
 
@@ -329,8 +339,9 @@ public class DbManager {
             ps.setInt(9, dailyPrepItem.getDailyPrepId());
             ps.executeUpdate();
             ps.close();
+            log.info("DailyPrepItem id:{} updated successfully", dailyPrepItem.getDailyPrepId());
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Failed to add DailyPrepItem: {}", e.getMessage());
         }
     }
 
@@ -365,8 +376,9 @@ public class DbManager {
                 }
                 return transaction;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to retrieve the last transaction: {}", e.getMessage());
             } finally {
+                log.info("Latest transaction retrieved successfully.");
                 ps.close();
                 rs.close();
             }
@@ -397,13 +409,13 @@ public class DbManager {
                 }
                 return dailyPreps;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to retrieve the dailyPrepData: {}", e.getMessage());
             } finally {
+                log.info("All DailyPredData entries retrieved.");
                 ps.close();
                 rs.close();
             }
         }
-        System.out.println("Get All Transactions failed! DB is not connected");
         return null;
     }
 
@@ -436,7 +448,7 @@ public class DbManager {
                 }
                 return dailyPrepItems;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Get daily prep item id: {} failed: {}", id, e.getMessage());
             }
         }
         return null;
@@ -471,9 +483,10 @@ public class DbManager {
                             rs.getDouble("hh_ll_any_high"),
                             rs.getDouble("hh_ll_any_low"));
                 }
+
                 return dailyPrepItem;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to retrieve dailyPrepItem id:{} : {}", id, e.getMessage());
             }
         }
         return null;
@@ -495,7 +508,7 @@ public class DbManager {
                 ps.close();
                 return getDailyPrepItem(id, symbol);
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to add dailyPrepItem with id: {} : {}", id, e.getMessage());
             }
         }
         return null;
@@ -511,9 +524,10 @@ public class DbManager {
                 ps.executeUpdate();
                 ps.close();
                 DailyPrepDate dailyPrepDate = getDailyPrepDate(date);
+                log.info("Daily Prep date added successfully.");
                 return new DailyPrep(dailyPrepDate.getDailyPrepDateId(), dailyPrepDate.getDate(), null, new ArrayList<>());
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to add dailyPrepDate with date: {} : {}", Date.valueOf(date), e.getMessage());
             }
         }
         return null;
@@ -534,8 +548,9 @@ public class DbManager {
             ps.close();
             return dailyPrepDate;
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Failed to retrieve dailyPrepDate with date: {} : {}", Date.valueOf(date), e.getMessage());
         }
+        return null;
     }
 
     public void deleteTransaction(Transaction transaction) throws SQLException {
@@ -547,8 +562,9 @@ public class DbManager {
                 preparedStatement.setInt(1, transaction.getId()); // Set the ID value
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Transaction id: {} deleted successfully.", transaction.getId());
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete transaction id: {} : {}", transaction.getId(), e.getMessage());
             }
         }
     }
@@ -562,8 +578,9 @@ public class DbManager {
                 preparedStatement.setInt(1, id); // Set the ID value
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Day id: {} deleted successfully.", id);
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete day id: {} : {}", id, e.getMessage());
             }
         }
     }
@@ -577,8 +594,9 @@ public class DbManager {
                 preparedStatement.setInt(1, id); // Set the ID value
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Symbol deleted successfully by dateid: {}", id);
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete symbol by dateid: {} : {}", id, e.getMessage());
             }
         }
     }
@@ -592,8 +610,9 @@ public class DbManager {
                 preparedStatement.setInt(1, id); // Set the ID value
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Symbol deleted successfully id: {}", id);
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete symbol by id: {} : {}", id, e.getMessage());
             }
         }
     }
@@ -617,13 +636,13 @@ public class DbManager {
                 }
                 return journalEntries;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to retrieve all journal entries : {}", e.getMessage());
             } finally {
+                log.info("All journal entries retrieved.");
                 ps.close();
                 rs.close();
             }
         }
-        System.out.println("Get All Journal Entries failed! DB is not connected");
         return null;
     }
 
@@ -638,8 +657,9 @@ public class DbManager {
                 ps.setString(3, journal.getText());
                 ps.executeUpdate();
                 ps.close();
+                log.info("Journal entry {} symbol {} added successfully.", Date.valueOf(journal.getDate()), journal.getSymbol());
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to add journal entry with date: {} : {}", journal.getDate(), e.getMessage());
             }
         }
     }
@@ -653,8 +673,9 @@ public class DbManager {
                 preparedStatement.setDate(1, Date.valueOf(date)); //
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Day: {} deleted successfully.", Date.valueOf(date));
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete journal by date: {} : {}", Date.valueOf(date), e.getMessage());
             }
         }
     }
@@ -668,8 +689,9 @@ public class DbManager {
                 preparedStatement.setInt(1, id); // Set the ID value
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Journal entry with symbol id {} deleted successfully.", id);
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete Journal entry by symbol id:{} : {}", id, e.getMessage());
             }
         }
     }
@@ -684,8 +706,9 @@ public class DbManager {
             ps.setInt(2, entry.getId());
             ps.executeUpdate();
             ps.close();
+            log.info("Journal entry id:{} updates successfully", entry.getId());
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Failed to update Journal entry with id: {} : {}", entry.getId(), e.getMessage());
         }
     }
 
@@ -709,13 +732,13 @@ public class DbManager {
                 }
                 return goals;
             } catch (Exception e) {
-                throw new RuntimeException(e);
+                log.error("Failed to retrieve all goals : {}", e.getMessage());
             } finally {
+                log.info("All goal entries retrieved.");
                 ps.close();
                 rs.close();
             }
         }
-        System.out.println("Get All Transactions failed! DB is not connected");
         return null;
     }
 
@@ -731,8 +754,9 @@ public class DbManager {
             ps.setBoolean(4, goal.getAchieved());
             ps.executeUpdate();
             ps.close();
+            log.info("Goal added successfully");
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Failed to add goal with date: {} : {}", goal.getDate(), e.getMessage());
         }
     }
 
@@ -745,8 +769,9 @@ public class DbManager {
                 preparedStatement.setInt(1, id); // Set the ID value
                 preparedStatement.executeUpdate();
                 preparedStatement.close();
+                log.info("Goal id: {} deleted successfully", id);
             } catch (SQLException e) {
-                e.printStackTrace();
+                log.error("Failed to delete goal witg id:{} : {}", id, e.getMessage());
             }
         }
     }
@@ -758,12 +783,13 @@ public class DbManager {
         try {
             ps = bdConnection.prepareStatement(query);
             ps.setString(1, goal.getText());
-            ps.setInt(2, goal.getId());
-            ps.setBoolean(3, goal.getAchieved());
+            ps.setBoolean(2, goal.getAchieved());
+            ps.setInt(3, goal.getId());
             ps.executeUpdate();
             ps.close();
+            log.info("Goal id: {} updated successfully", goal.getId());
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            log.error("Failed to update goal id:{} : {}", goal.getId(), e.getMessage());
         }
     }